+++ packages/zwave-js/src/lib/driver/SendThreadMachine.test.ts
@@ -32,6 +32,7 @@ import type { Driver } from "./Driver";
 import {
 	createSendThreadMachine,
 	SendThreadInterpreter,
+	SendThreadMachineParams,
 } from "./SendThreadMachine";
 import { Transaction } from "./Transaction";
 
@@ -120,6 +121,20 @@ mockCommandQueueMachine.mockReturnValue(
 		},
 	}),
 );
+
+const machineParams: SendThreadMachineParams = {
+	timeouts: {
+		ack: 1000,
+		response: 1600,
+		report: 1600,
+		sendDataCallback: 65000,
+	},
+	attempts: {
+		controller: 3,
+		// @ts-expect-error this property is not defined, we just use it to deduplicate test code
+		sendData: 3,
+	},
+};
 
 describe("lib/driver/SendThreadMachine", () => {
 	beforeEach(() => {
@@ -423,7 +438,9 @@ describe("lib/driver/SendThreadMachine", () => {
 		{
 			guards: {
 				expectsUpdate: (ctx) => !!ctx.expectsUpdate,
-				mayRetry: (ctx) => ctx.sendDataAttempts < 3,
+				mayRetry: (ctx) =>
+					ctx.sendDataAttempts <
+					(machineParams.attempts as any).sendData,
 			},
 		},
 	);
@@ -658,6 +675,7 @@ describe("lib/driver/SendThreadMachine", () => {
 
 				it(path.description, () => {
 					const fakeDriver = (createEmptyMockDriver() as unknown) as Driver;
+					fakeDriver.options.attempts.sendData = (machineParams.attempts as any).sendData;
 					const sm = new SecurityManager({
 						ownNodeId: 1,
 						nonceTimeout: 500,
@@ -776,6 +794,7 @@ describe("lib/driver/SendThreadMachine", () => {
 					};
 					const machine = createSendThreadMachine(
 						implementations as any,
+						machineParams,
 					);
 
 					const context: TestContext = {